% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factoid_table.R
\name{factoid_table}
\alias{factoid_table}
\title{Build a factoid table.}
\usage{
factoid_table(year, us_pres, best_pic, ticket_price)
}
\arguments{
\item{ticket_price}{}
}
\value{

}
\description{
Build a factoid table.
}
