% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_posters.R
\name{plot_posters}
\alias{plot_posters}
\title{Plot a row of movie posters}
\usage{
plot_posters(img_files, theme_size = 6)
}
\arguments{
\item{theme_size}{}
}
\value{
ggplot2 image
}
\description{
Plot a row of movie posters
}
